tests/day08.test.ts
import { navigateStorm, navigateStormWithGhosts } from '../src/day08';

const input1 = `RL

AAA = (BBB, CCC)
BBB = (DDD, EEE)
CCC = (ZZZ, GGG)
DDD = (DDD, DDD)
EEE = (EEE, EEE)
GGG = (GGG, GGG)
ZZZ = (ZZZ, ZZZ)`;

const input2 = `LLR

AAA = (BBB, BBB)
BBB = (AAA, ZZZ)
ZZZ = (ZZZ, ZZZ)`;

test('day 8-1', () => {
	expect(navigateStorm(input1)).toBe(2);
	expect(navigateStorm(input2)).toBe(6);
});

const input3 = `LR

11A = (11B, XXX)
11B = (XXX, 11Z)
11Z = (11B, XXX)
22A = (22B, XXX)
22B = (22C, 22C)
22C = (22Z, 22Z)
22Z = (22B, 22B)
XXX = (XXX, XXX)`;

test('day 8-2', () => {
	expect(navigateStormWithGhosts(input3)).toBe(6);
});
